requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Serial communication
pyserial==3.5

# MQTT
paho-mqtt==2.0.0

# Async support
asyncio-mqtt==0.16.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# HTML templates
jinja2==3.1.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
